fix(salud): el trinquete tenía razón · dos regresiones que traía el PR anterior
CI en rojo por el marcador de salud, y las dos causas eran mías.

── todos_totales 231 → 233 ────────────────────────────────────────────────

Escribí «TODO» en mayúsculas como énfasis en dos comentarios en español —«con
TODO lo adelantado», «manda sobre TODO lo de abajo»— y el detector cuenta
`\b(TODO|FIXME|HACK|XXX)\b` como deuda-marcador. Desde fuera son
indistinguibles de un TODO real, así que el énfasis pasa a **negrita**, que es
como lo marca el resto de la casa.

── archivos_800 38 → 39 ───────────────────────────────────────────────────

`types/prestamos.ts` estaba en 799 líneas, a UNA del tope, y el campo
`esAdelantoDeCapital` con su explicación lo dejó en 815. Recortar el comentario
para pasar por debajo habría dejado el fichero clavado en el límite y al
siguiente que tocara el cuadro se le pondría CI en rojo sin saber por qué: es
una trampa, no un arreglo.

Así que el cuadro se va a `types/planPagos.ts` —`PeriodoPago`, `PlanPagos` y
`CalculoAmortizacion`— y `types/prestamos.ts` los reexporta, con lo que ningún
import de fuera cambia. El corte es por tema y no por la tijera: un préstamo es
lo que se pactó —tipo, plazo, comisiones, bonificaciones—, y esto es lo que sale
de aplicarlo, recibo a recibo. El fichero queda en 763 líneas, con margen real.

`ficheros_no_v5` 109 → 110 sale con ⇧amp: es la exención ya registrada de #1643,
no la trae este PR.

Trinquete OK en local, `tsc` limpio y la suite completa deja exactamente los
mismos fallos que la rama base.
